Move paragraphs legacy tests
